requirements.txt
selenium>=4.23.1
webdriver-manager>=4.0.2
pytz>=2024.1
google-generativeai>=0.8.0
